requirements.txt
accelerate==0.24.1
diffusers==0.23.0
gradio==3.39.0
invisible-watermark==0.2.0
Pillow==10.1.0
torch==2.1.0
transformers==4.35.0
